simplify proxying model, accomidating multiple scopes.
